salva os desenhos no firestore
